% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.escape_regex}
\alias{.escape_regex}
\title{Escapes any characters that would have special
meaning in a regular expression}
\usage{
.escape_regex(string)
}
\arguments{
\item{string}{(\code{string})\cr
A character vector.}
}
\description{
This functions has been adapted from \code{Hmisc::escapeRegex()}
}
\seealso{
Other other_helpers:
\code{\link[=.clean_backticks]{.clean_backticks()}}
}
\concept{other_helpers}
