% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_utilities.R
\name{.generic_selector}
\alias{.generic_selector}
\alias{.is_selector_scoped}
\title{Generate a custom selector function}
\usage{
.generic_selector(variable_column, select_column, select_expr, fun_name)

.is_selector_scoped(variable_column, select_column)
}
\arguments{
\item{variable_column}{string indicating column variable names are stored}

\item{select_column}{character vector of columns used in the \verb{select_expr=} argument}

\item{select_expr}{unquoted predicate command to subset a data frame to select variables}

\item{fun_name}{quoted name of function where \code{.generic_selector()} is being used.
This helps with error messaging.}
}
\value{
custom selector functions
}
\description{
Generate a custom selector function
}
\details{
\code{.is_selector_scoped()} checks if a selector has been properly registered
in \code{env_variable_type$df_var_info}.
}
