% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_get_pairwise_contrasts.R
\name{model_get_pairwise_contrasts}
\alias{model_get_pairwise_contrasts}
\title{Get pairwise comparison of the levels of a categorical variable}
\usage{
model_get_pairwise_contrasts(
  model,
  variables,
  pairwise_reverse = TRUE,
  contrasts_adjust = NULL,
  conf.level = 0.95,
  emmeans_args = list()
)
}
\arguments{
\item{model}{(a model object, e.g. \code{glm})\cr
A model object.}

\item{variables}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
Variables to add pairwise contrasts.}

\item{pairwise_reverse}{(\code{logical})\cr
Determines whether to use \code{"pairwise"} (if \code{TRUE})
or \code{"revpairwise"} (if \code{FALSE}), see \code{\link[emmeans:contrast]{emmeans::contrast()}}.}

\item{contrasts_adjust}{optional adjustment method when computing contrasts,
see \code{\link[emmeans:contrast]{emmeans::contrast()}} (if \code{NULL}, use \code{emmeans} default)}

\item{conf.level}{(\code{numeric})\cr
Level of confidence for confidence intervals (default: 95\%).}

\item{emmeans_args}{(\code{logical})\cr
List of additional parameter to pass to
\code{\link[emmeans:emmeans]{emmeans::emmeans()}} when computing pairwise contrasts.}
}
\description{
It is computed with \code{\link[emmeans:emmeans]{emmeans::emmeans()}}.
}
\details{
For \code{pscl::zeroinfl()} and \code{pscl::hurdle()} models, pairwise contrasts are
computed separately for each component, using \code{mode = "count"} and
\code{mode = "zero"} (see documentation of \code{emmeans}) and a component column
is added to the results.
}
\examples{
\dontshow{if (.assert_package("emmeans", boolean = TRUE)) withAutoprint(\{ # examplesIf}
\donttest{
  mod <- lm(Sepal.Length ~ Species, data = iris)
  mod |> model_get_pairwise_contrasts(variables = "Species")
  mod |>
    model_get_pairwise_contrasts(
      variables = "Species",
      contrasts_adjust = "none"
    )
}
\dontshow{\}) # examplesIf}
}
\seealso{
Other model_helpers:
\code{\link[=model_compute_terms_contributions]{model_compute_terms_contributions()}},
\code{\link[=model_get_assign]{model_get_assign()}},
\code{\link[=model_get_coefficients_type]{model_get_coefficients_type()}},
\code{\link[=model_get_contrasts]{model_get_contrasts()}},
\code{\link[=model_get_model]{model_get_model()}},
\code{\link[=model_get_model_frame]{model_get_model_frame()}},
\code{\link[=model_get_model_matrix]{model_get_model_matrix()}},
\code{\link[=model_get_n]{model_get_n()}},
\code{\link[=model_get_nlevels]{model_get_nlevels()}},
\code{\link[=model_get_offset]{model_get_offset()}},
\code{\link[=model_get_response]{model_get_response()}},
\code{\link[=model_get_response_variable]{model_get_response_variable()}},
\code{\link[=model_get_terms]{model_get_terms()}},
\code{\link[=model_get_weights]{model_get_weights()}},
\code{\link[=model_get_xlevels]{model_get_xlevels()}},
\code{\link[=model_identify_variables]{model_identify_variables()}},
\code{\link[=model_list_contrasts]{model_list_contrasts()}},
\code{\link[=model_list_higher_order_variables]{model_list_higher_order_variables()}},
\code{\link[=model_list_terms_levels]{model_list_terms_levels()}},
\code{\link[=model_list_variables]{model_list_variables()}}
}
\concept{model_helpers}
