% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_add_header_rows.R
\name{tidy_add_header_rows}
\alias{tidy_add_header_rows}
\title{Add header rows variables with several terms}
\usage{
tidy_add_header_rows(
  x,
  show_single_row = NULL,
  model = tidy_get_model(x),
  quiet = FALSE,
  strict = FALSE
)
}
\arguments{
\item{x}{(\code{data.frame})\cr
A tidy tibble as produced by \verb{tidy_*()} functions.}

\item{show_single_row}{(\code{\link[dplyr:dplyr_tidy_select]{tidy-select}})\cr
Names of dichotomous variables that should be displayed on a single row.
See also \code{\link[=all_dichotomous]{all_dichotomous()}}.}

\item{model}{(a model object, e.g. \code{glm})\cr
The corresponding model, if not attached to \code{x}.}

\item{quiet}{(\code{logical})\cr
Whether \code{broom.helpers} should not return a message when requested output
cannot be generated. Default is \code{FALSE}.}

\item{strict}{(\code{logical})\cr
Whether \code{broom.helpers} should return an error
when requested output cannot be generated. Default is \code{FALSE}.}
}
\description{
For variables with several terms (usually categorical variables but
could also be the case of continuous variables with polynomial terms
or splines), \code{tidy_add_header_rows()} will add an additional row
per variable, where \code{label} will be equal to \code{var_label}.
These additional rows could be identified with \code{header_row} column.
}
\details{
The \code{show_single_row} argument allows to specify a list
of dichotomous variables that should be displayed on a single row
instead of two rows.

The added \code{header_row} column will be equal to:
\itemize{
\item \code{TRUE} for an header row;
\item \code{FALSE} for a normal row of a variable with an header row;
\item \code{NA} for variables without an header row.
}

If the \code{label} column is not yet available in \code{x},
\code{\link[=tidy_add_term_labels]{tidy_add_term_labels()}} will be automatically applied.
}
\examples{
\dontshow{if (.assert_package("gtsummary", boolean = TRUE)) withAutoprint(\{ # examplesIf}
\donttest{
  df <- Titanic |>
    dplyr::as_tibble() |>
    dplyr::mutate(Survived = factor(Survived, c("No", "Yes")))

  res <-
    glm(
      Survived ~ Class + Age + Sex,
      data = df, weights = df$n, family = binomial,
      contrasts = list(Age = contr.sum, Class = "contr.SAS")
    ) |>
    tidy_and_attach() |>
    tidy_add_variable_labels(labels = list(Class = "Custom label for Class")) |>
    tidy_add_reference_rows()
  res |> tidy_add_header_rows()
  res |> tidy_add_header_rows(show_single_row = all_dichotomous())

  glm(
    response ~ stage + grade * trt,
    gtsummary::trial,
    family = binomial,
    contrasts = list(
      stage = contr.treatment(4, base = 3),
      grade = contr.treatment(3, base = 2),
      trt = contr.treatment(2, base = 2)
    )
  ) |>
    tidy_and_attach() |>
    tidy_add_reference_rows() |>
    tidy_add_header_rows()
}
\dontshow{\}) # examplesIf}
}
\seealso{
Other tidy_helpers:
\code{\link[=tidy_add_coefficients_type]{tidy_add_coefficients_type()}},
\code{\link[=tidy_add_contrasts]{tidy_add_contrasts()}},
\code{\link[=tidy_add_estimate_to_reference_rows]{tidy_add_estimate_to_reference_rows()}},
\code{\link[=tidy_add_n]{tidy_add_n()}},
\code{\link[=tidy_add_pairwise_contrasts]{tidy_add_pairwise_contrasts()}},
\code{\link[=tidy_add_reference_rows]{tidy_add_reference_rows()}},
\code{\link[=tidy_add_term_labels]{tidy_add_term_labels()}},
\code{\link[=tidy_add_variable_labels]{tidy_add_variable_labels()}},
\code{\link[=tidy_attach_model]{tidy_attach_model()}},
\code{\link[=tidy_disambiguate_terms]{tidy_disambiguate_terms()}},
\code{\link[=tidy_group_by]{tidy_group_by()}},
\code{\link[=tidy_identify_variables]{tidy_identify_variables()}},
\code{\link[=tidy_plus_plus]{tidy_plus_plus()}},
\code{\link[=tidy_remove_intercept]{tidy_remove_intercept()}},
\code{\link[=tidy_select_variables]{tidy_select_variables()}}
}
\concept{tidy_helpers}
