% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tidiers.R
\name{tidy_broom}
\alias{tidy_broom}
\title{Tidy with \code{broom::tidy()} and checks that all arguments are used}
\usage{
tidy_broom(x, ...)
}
\arguments{
\item{x}{(a model object, e.g. \code{glm})\cr
A model to be tidied.}

\item{...}{Additional parameters passed to \code{broom::tidy()}.}
}
\description{
Tidy with \code{broom::tidy()} and checks that all arguments are used
}
\seealso{
Other custom_tidiers:
\code{\link[=tidy_coxphms]{tidy_coxphms()}},
\code{\link[=tidy_multgee]{tidy_multgee()}},
\code{\link[=tidy_parameters]{tidy_parameters()}},
\code{\link[=tidy_svy_vglm]{tidy_svy_vglm()}},
\code{\link[=tidy_vgam]{tidy_vgam()}},
\code{\link[=tidy_with_broom_or_parameters]{tidy_with_broom_or_parameters()}},
\code{\link[=tidy_zeroinfl]{tidy_zeroinfl()}}
}
\concept{custom_tidiers}
