% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_tidiers.R
\name{tidy_coxphms}
\alias{tidy_coxphms}
\title{Tidy a multi-state survival model}
\usage{
tidy_coxphms(x, conf.int = TRUE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{(\code{coxphms})\cr
A \code{survival::coxphms.object} model.}

\item{conf.int}{(\code{logical})\cr
Whether or not to include a confidence interval in the tidied output.}

\item{conf.level}{(\code{numeric})\cr
The confidence level to use for the confidence interval
(between \code{0} ans \code{1}).}

\item{...}{Additional parameters passed to \code{parameters::model_parameters()}.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
A tidier for multi-state models (\code{survival::coxphms.object}) generated
with \code{survival::coxph()}.
Term names will be updated to be consistent with generic models. The original
term names are preserved in an \code{"original_term"} column. An additional column
\code{"state"} will be added to provide status detail (i.e. the right part of
original terms) and a column \code{"y.level"} will be populated by using values
stored in \code{x$states}.
}
\seealso{
Other custom_tidiers: 
\code{\link{tidy_broom}()},
\code{\link{tidy_multgee}()},
\code{\link{tidy_parameters}()},
\code{\link{tidy_svy_vglm}()},
\code{\link{tidy_vgam}()},
\code{\link{tidy_with_broom_or_parameters}()},
\code{\link{tidy_zeroinfl}()}
}
\concept{custom_tidiers}
