% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_tidiers.R
\name{tidy_ggpredict}
\alias{tidy_ggpredict}
\title{Marginal Predictions with \code{ggeffects::ggpredict()}}
\usage{
tidy_ggpredict(x, conf.int = TRUE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{(a model object, e.g. \code{glm})\cr
A model to be tidied.}

\item{conf.int}{(\code{logical})\cr
Whether or not to include a confidence interval in the tidied output.}

\item{conf.level}{(\code{numeric})\cr
The confidence level to use for the confidence interval (between \code{0} ans \code{1}).}

\item{...}{Additional parameters passed to \code{ggeffects::ggpredict()}.}
}
\description{
Use \code{ggeffects::ggpredict()} to estimate marginal predictions
and return a tibble tidied in a way that it could be used by \code{broom.helpers}
functions.
See \url{https://strengejacke.github.io/ggeffects/} for a list of supported
models.
}
\details{
By default, \code{ggeffects::ggpredict()} estimate marginal predictions at the
observed mean of continuous variables and at the first modality of categorical
variables (regardless of the type of contrasts used in the model).

For more information, see \code{vignette("marginal_tidiers", "broom.helpers")}.
}
\note{
By default, \code{ggeffects::ggpredict()} estimates marginal predictions for each
individual variable, regardless of eventual interactions.
}
\examples{
\dontshow{if (.assert_package("ggeffects", boolean = TRUE)) withAutoprint(\{ # examplesIf}
\donttest{
df <- Titanic |>
  dplyr::as_tibble() |>
  tidyr::uncount(n) |>
  dplyr::mutate(Survived = factor(Survived, c("No", "Yes")))
mod <- glm(
  Survived ~ Class + Age + Sex,
  data = df, family = binomial
)
tidy_ggpredict(mod)
tidy_plus_plus(mod, tidy_fun = tidy_ggpredict)
}
\dontshow{\}) # examplesIf}
}
\seealso{
\code{ggeffects::ggpredict()}

Other marginal_tieders:
\code{\link[=tidy_all_effects]{tidy_all_effects()}},
\code{\link[=tidy_avg_comparisons]{tidy_avg_comparisons()}},
\code{\link[=tidy_avg_slopes]{tidy_avg_slopes()}},
\code{\link[=tidy_marginal_contrasts]{tidy_marginal_contrasts()}},
\code{\link[=tidy_marginal_predictions]{tidy_marginal_predictions()}},
\code{\link[=tidy_margins]{tidy_margins()}}
}
\concept{marginal_tieders}
