% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_tidiers.R
\name{tidy_marginal_means}
\alias{tidy_marginal_means}
\title{Marginal Means with deprecated \code{marginaleffects::marginal_means()}}
\usage{
tidy_marginal_means(x, conf.int = TRUE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{(a model object, e.g. \code{glm})\cr
A model to be tidied.}

\item{conf.int}{(\code{logical})\cr
Whether or not to include a confidence interval in the tidied output.}

\item{conf.level}{(\code{numeric})\cr
The confidence level to use for the confidence interval (between \code{0} ans \code{1}).}

\item{...}{Additional parameters.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
This function is deprecated. \code{marginal_means()} is not anymore exported
by \code{marginaleffects}. Use instead \code{tidy_marginal_predictions()} with
the option \code{newdata = "marginalmeans"}.
}
