% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_tidiers.R
\name{tidy_marginalmeans}
\alias{tidy_marginalmeans}
\title{Marginal Means with \code{marginaleffects::marginalmeans()}}
\usage{
tidy_marginalmeans(x, conf.int = TRUE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{a model}

\item{conf.int}{logical indicating whether or not to include a confidence
interval in the tidied output}

\item{conf.level}{the confidence level to use for the confidence interval}

\item{...}{additional parameters passed to
\code{marginaleffects::marginalmeans()}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Use \code{marginaleffects::marginalmeans()} to estimate marginal means and
return a tibble tidied in a way that it could be used by \code{broom.helpers}
functions. See \code{marginaleffects::marginalmeans()()} for a list of supported
models.
}
\details{
\code{marginaleffects::marginalmeans()} estimate marginal means:
adjusted predictions, averaged across a grid of categorical predictors,
holding other numeric predictors at their means. Please refer to the
documentation page of \code{marginaleffects::marginalmeans()}. Marginal means
are defined only for categorical variables.

For more information, see \code{vignette("marginal_tidiers", "broom.helpers")}.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Average Marginal Means

df <- Titanic \%>\%
  dplyr::as_tibble() \%>\%
  tidyr::uncount(n) \%>\%
  dplyr::mutate(Survived = factor(Survived, c("No", "Yes")))
mod <- glm(
  Survived ~ Class + Age + Sex,
  data = df, family = binomial
)
tidy_marginalmeans(mod)
tidy_plus_plus(mod, tidy_fun = tidy_marginalmeans)

mod2 <- lm(Petal.Length ~ poly(Petal.Width, 2) + Species, data = iris)
tidy_marginalmeans(mod2)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{marginaleffects::marginalmeans()}

Other marginal_tieders: 
\code{\link{tidy_all_effects}()},
\code{\link{tidy_comparisons}()},
\code{\link{tidy_ggpredict}()},
\code{\link{tidy_marginalcontrasts}()},
\code{\link{tidy_marginaleffects}()},
\code{\link{tidy_marginalpredictions}()},
\code{\link{tidy_margins}()}
}
\concept{marginal_tieders}
