% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_tidiers.R
\name{tidy_marginalpredictions}
\alias{tidy_marginalpredictions}
\alias{variables_to_predict}
\alias{plot_marginalpredictions}
\title{Marginal Predictions with \code{marginaleffects::predictions()}}
\usage{
tidy_marginalpredictions(
  x,
  variables_list = "auto",
  conf.int = TRUE,
  conf.level = 0.95,
  ...
)

variables_to_predict(
  model,
  interactions = TRUE,
  categorical = unique,
  continuous = "fivenum"
)

plot_marginalpredictions(x, variables_list = "auto", conf.level = 0.95, ...)
}
\arguments{
\item{x}{a model}

\item{variables_list}{a list whose elements will be sequentially passed to
\code{variables} in \code{marginaleffects::predictions()} (see details below);
alternatively, it could also be the string \code{"auto"} (default) or
\code{"no_interaction"}}

\item{conf.int}{logical indicating whether or not to include a confidence
interval in the tidied output}

\item{conf.level}{the confidence level to use for the confidence interval}

\item{...}{additional parameters passed to
\code{marginaleffects::predictions()}}

\item{model}{a model}

\item{interactions}{should combinations of variables corresponding to
interactions be returned?}

\item{categorical}{default value for categorical variables}

\item{continuous}{default value for continuous variables}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Use \code{marginaleffects::predictions()} to estimate marginal predictions for
each variable of a model and return a tibble tidied in a way that it could
be used by \code{broom.helpers} functions.
See \code{marginaleffects::predictions()} for a list of supported models.
}
\details{
Marginal predictions are obtained by calling, for each variable,
\code{marginaleffects::predictions()} with the same variable being used for the
\code{variables} and the \code{by} argument.

Considering a categorical variable named \code{cat}, \code{tidy_marginalpredictions()}
will call \code{predictions(model, variables = list(cat = unique), by = "cat")} to
obtain average marginal predictions for this variable.

Considering a continuous variable named \code{cont}, \code{tidy_marginalpredictions()}
will call \code{predictions(model, variables = list(cont = "fivenum"), by = "cont")}
to obtain average marginal predictions for this variable at the minimum, the
first quartile, the median, the third quartile and the maximum of the observed
values of \code{cont}.

By default, \emph{average marginal predictions} are computed: predictions are made
using a counterfactual grid for each value of the variable of interest,
before averaging the results. \emph{Marginal predictions at the mean} could be
obtained by indicating \code{newdata = "mean"}. Other assumptions are possible,
see the help file of \code{marginaleffects::predictions()}.

\code{tidy_marginalpredictions()} will compute marginal predictions for each
variable or combination of variables, before stacking the results in a unique
tibble. This is why \code{tidy_marginalpredictions()} has a \code{variables_list}
argument consisting of a list of specifications that will be passed
sequentially to the \code{variables} argument of \code{marginaleffects::predictions()}.

The helper function \code{variables_to_predict()} could be used to automatically
generate a suitable list to be used with \code{variables_list}. By default, all
unique values are retained for categorical variables and \code{"fivenum"} (i.e.
Tukey's five numbers, minimum, quartiles and maximum) for continuous variables.
When \code{interactions = FALSE}, \code{variables_to_predict()} will return a list of
all individual variables used in the model. If \code{interactions = FALSE}, it
will search for higher order combinations of variables (see
\code{model_list_higher_order_variables()}).

\code{variables_list}'s default value, \code{"auto"}, calls
\code{variables_to_predict(interactions = TRUE)} while \code{"no_interaction"} is a
shortcut for \code{variables_to_predict(interactions = FALSE)}.

You can also provide custom specifications (see examples).

\code{plot_marginalpredictions()} works in a similar way and returns a list of
plots that could be combined with \code{patchwork::wrap_plots()} (see examples).

For more information, see \code{vignette("marginal_tidiers", "broom.helpers")}.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Average Marginal Predictions
df <- Titanic \%>\%
  dplyr::as_tibble() \%>\%
  tidyr::uncount(n) \%>\%
  dplyr::mutate(Survived = factor(Survived, c("No", "Yes")))
mod <- glm(
  Survived ~ Class + Age + Sex,
  data = df, family = binomial
)
tidy_marginalpredictions(mod)
tidy_plus_plus(mod, tidy_fun = tidy_marginalpredictions)
plot_marginalpredictions(mod) \%>\% patchwork::wrap_plots()
plot_marginalpredictions(mod) \%>\%
  patchwork::wrap_plots() &
  ggplot2::scale_y_continuous(limits = c(0, 1), label = scales::percent)

mod2 <- lm(Petal.Length ~ poly(Petal.Width, 2) + Species, data = iris)
tidy_marginalpredictions(mod2)
plot_marginalpredictions(mod2) \%>\% patchwork::wrap_plots()
tidy_marginalpredictions(
  mod2,
  variables_list = variables_to_predict(mod2, continuous = "threenum")
)
tidy_marginalpredictions(
  mod2,
  variables_list = list(
    list(Petal.Width = c(0, 1, 2, 3)),
    list(Species = unique)
  )
)
tidy_marginalpredictions(
  mod2,
  variables_list = list(list(Species = unique, Petal.Width = 1:3))
)

# Model with interactions
mod3 <- glm(
  Survived ~ Sex * Age + Class,
  data = df, family = binomial
)
tidy_marginalpredictions(mod3)
plot_marginalpredictions(mod3) \%>\% patchwork::wrap_plots()
tidy_marginalpredictions(mod3, "no_interaction")
plot_marginalpredictions(mod3, "no_interaction") \%>\% patchwork::wrap_plots()
tidy_marginalpredictions(
  mod3,
  variables_list = list(
    list(Class = unique, Sex = "Female"),
    list(Age = unique)
  )
)

# Marginal Predictions at the Mean
tidy_marginalpredictions(mod, newdata = "mean")
plot_marginalpredictions(mod, newdata = "mean") \%>\% patchwork::wrap_plots()
\dontshow{\}) # examplesIf}
}
\seealso{
\code{marginaleffects::predictions()}

Other marginal_tieders: 
\code{\link{tidy_all_effects}()},
\code{\link{tidy_comparisons}()},
\code{\link{tidy_ggpredict}()},
\code{\link{tidy_marginalcontrasts}()},
\code{\link{tidy_marginaleffects}()},
\code{\link{tidy_marginalmeans}()},
\code{\link{tidy_margins}()}
}
\concept{marginal_tieders}
