% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{.clean_backticks}
\alias{.clean_backticks}
\title{Remove backtips around variable names}
\usage{
.clean_backticks(x, variable_names = x)
}
\arguments{
\item{x}{a character vector of variable names to be cleaned}

\item{variable_names}{optional character vector of possible
variable names, for example obtained with
\link[=model_list_variables]{model_list_variables(only_variable = TRUE)}}
}
\description{
Remove backtips around variable names
}
\seealso{
Other other_helpers: 
\code{\link{.escape_regex}()}
}
\concept{other_helpers}
