% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_get_coefficients_type.R
\name{model_get_coefficients_type}
\alias{model_get_coefficients_type}
\alias{model_get_coefficients_type.default}
\alias{model_get_coefficients_type.glm}
\alias{model_get_coefficients_type.negbin}
\alias{model_get_coefficients_type.geeglm}
\alias{model_get_coefficients_type.fixest}
\alias{model_get_coefficients_type.biglm}
\alias{model_get_coefficients_type.glmerMod}
\alias{model_get_coefficients_type.clogit}
\alias{model_get_coefficients_type.polr}
\alias{model_get_coefficients_type.multinom}
\alias{model_get_coefficients_type.svyolr}
\alias{model_get_coefficients_type.clm}
\alias{model_get_coefficients_type.clmm}
\alias{model_get_coefficients_type.coxph}
\alias{model_get_coefficients_type.crr}
\alias{model_get_coefficients_type.tidycrr}
\alias{model_get_coefficients_type.cch}
\alias{model_get_coefficients_type.model_fit}
\alias{model_get_coefficients_type.LORgee}
\alias{model_get_coefficients_type.vglm}
\alias{model_get_coefficients_type.vgam}
\alias{model_get_coefficients_type.svy_vglm}
\alias{model_get_coefficients_type.brmsfit}
\title{Get coefficient type}
\usage{
model_get_coefficients_type(model)

\method{model_get_coefficients_type}{default}(model)

\method{model_get_coefficients_type}{glm}(model)

\method{model_get_coefficients_type}{negbin}(model)

\method{model_get_coefficients_type}{geeglm}(model)

\method{model_get_coefficients_type}{fixest}(model)

\method{model_get_coefficients_type}{biglm}(model)

\method{model_get_coefficients_type}{glmerMod}(model)

\method{model_get_coefficients_type}{clogit}(model)

\method{model_get_coefficients_type}{polr}(model)

\method{model_get_coefficients_type}{multinom}(model)

\method{model_get_coefficients_type}{svyolr}(model)

\method{model_get_coefficients_type}{clm}(model)

\method{model_get_coefficients_type}{clmm}(model)

\method{model_get_coefficients_type}{coxph}(model)

\method{model_get_coefficients_type}{crr}(model)

\method{model_get_coefficients_type}{tidycrr}(model)

\method{model_get_coefficients_type}{cch}(model)

\method{model_get_coefficients_type}{model_fit}(model)

\method{model_get_coefficients_type}{LORgee}(model)

\method{model_get_coefficients_type}{vglm}(model)

\method{model_get_coefficients_type}{vgam}(model)

\method{model_get_coefficients_type}{svy_vglm}(model)

\method{model_get_coefficients_type}{brmsfit}(model)
}
\arguments{
\item{model}{(a model object, e.g. \code{glm})\cr
A model object.}
}
\description{
Indicate the type of coefficient among "generic", "logistic",
"poisson", "relative_risk" or "prop_hazard".
}
\examples{
lm(hp ~ mpg + factor(cyl), mtcars) |>
  model_get_coefficients_type()

df <- Titanic |>
  dplyr::as_tibble() |>
  dplyr::mutate(Survived = factor(Survived, c("No", "Yes")))
glm(Survived ~ Class + Age * Sex, data = df, weights = df$n, family = binomial) |>
  model_get_coefficients_type()
}
\seealso{
Other model_helpers:
\code{\link[=model_compute_terms_contributions]{model_compute_terms_contributions()}},
\code{\link[=model_get_assign]{model_get_assign()}},
\code{\link[=model_get_contrasts]{model_get_contrasts()}},
\code{\link[=model_get_model]{model_get_model()}},
\code{\link[=model_get_model_frame]{model_get_model_frame()}},
\code{\link[=model_get_model_matrix]{model_get_model_matrix()}},
\code{\link[=model_get_n]{model_get_n()}},
\code{\link[=model_get_nlevels]{model_get_nlevels()}},
\code{\link[=model_get_offset]{model_get_offset()}},
\code{\link[=model_get_pairwise_contrasts]{model_get_pairwise_contrasts()}},
\code{\link[=model_get_response]{model_get_response()}},
\code{\link[=model_get_response_variable]{model_get_response_variable()}},
\code{\link[=model_get_terms]{model_get_terms()}},
\code{\link[=model_get_weights]{model_get_weights()}},
\code{\link[=model_get_xlevels]{model_get_xlevels()}},
\code{\link[=model_identify_variables]{model_identify_variables()}},
\code{\link[=model_list_contrasts]{model_list_contrasts()}},
\code{\link[=model_list_higher_order_variables]{model_list_higher_order_variables()}},
\code{\link[=model_list_terms_levels]{model_list_terms_levels()}},
\code{\link[=model_list_variables]{model_list_variables()}}
}
\concept{model_helpers}
