% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_get_contrasts.R
\name{model_get_contrasts}
\alias{model_get_contrasts}
\alias{model_get_contrasts.default}
\title{Get contrasts used in the model}
\usage{
model_get_contrasts(model)

\method{model_get_contrasts}{default}(model)
}
\arguments{
\item{model}{a model object}
}
\description{
Get contrasts used in the model
}
\seealso{
Other model_helpers: 
\code{\link{model_get_model_frame}()},
\code{\link{model_get_model_matrix}()},
\code{\link{model_get_xlevels}()},
\code{\link{model_identify_variables}()}
}
\concept{model_helpers}
