% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_get_offset.R
\name{model_get_offset}
\alias{model_get_offset}
\alias{model_get_offset.default}
\title{Get model offset}
\usage{
model_get_offset(model)

\method{model_get_offset}{default}(model)
}
\arguments{
\item{model}{a model object}
}
\description{
\lifecycle{experimental}
}
\examples{
mod <- glm(response ~ trt + offset(ttdeath), gtsummary::trial, family = poisson)
mod \%>\% model_get_offset()
}
\seealso{
Other model_helpers: 
\code{\link{model_compute_terms_contributions}()},
\code{\link{model_get_coefficients_type}()},
\code{\link{model_get_contrasts}()},
\code{\link{model_get_model_frame}()},
\code{\link{model_get_model_matrix}()},
\code{\link{model_get_model}()},
\code{\link{model_get_nlevels}()},
\code{\link{model_get_n}()},
\code{\link{model_get_response}()},
\code{\link{model_get_weights}()},
\code{\link{model_get_xlevels}()},
\code{\link{model_identify_variables}()},
\code{\link{model_list_contrasts}()},
\code{\link{model_list_terms_levels}()},
\code{\link{model_list_variables}()}
}
\concept{model_helpers}
