% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_helpers.R
\name{select_helpers}
\alias{select_helpers}
\alias{all_continuous}
\alias{all_dichotomous}
\alias{all_categorical}
\alias{all_interaction}
\alias{all_intercepts}
\alias{all_treatment_contrasts}
\alias{all_sum_contrasts}
\alias{all_poly_contrasts}
\alias{all_helmert_contrasts}
\title{Select helper functions}
\usage{
all_continuous()

all_dichotomous()

all_categorical(dichotomous = TRUE)

all_interaction()

all_intercepts()

all_treatment_contrasts()

all_sum_contrasts()

all_poly_contrasts()

all_helmert_contrasts()
}
\arguments{
\item{dichotomous}{Logical indicating whether to include dichotomous variables.
Default is \code{TRUE}}
}
\value{
A character vector of column names selected
}
\description{
Set of functions to supplement the {tidyselect} set of functions for selecting
columns of data frames.
}
\examples{
mod <- glm(response ~ age * trt + grade, gtsummary::trial, family = binomial)
res <- mod \%>\% tidy_plus_plus(exponentiate = TRUE, keep_model = TRUE)
res \%>\% tidy_select_variables(all_continuous())
res \%>\% tidy_select_variables(all_dichotomous())
res \%>\% tidy_select_variables(all_categorical())
res \%>\% tidy_select_variables(all_categorical(dichotomous = FALSE))
res \%>\% tidy_select_variables(all_interaction())
}
