% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_add_contrasts.R
\name{tidy_add_contrasts}
\alias{tidy_add_contrasts}
\title{Add contrasts type for categorical variables}
\usage{
tidy_add_contrasts(x, model = tidy_get_model(x))
}
\arguments{
\item{x}{a tidy tibble}

\item{model}{the corresponding model, if not attached to \code{x}}
}
\description{
Add a \code{contrasts} column corresponding to the type of contrasts
}
\details{
If the \code{variable} column is not yet available in \code{x},
\code{\link[=tidy_identify_variables]{tidy_identify_variables()}} will be automatically applied.
}
\examples{
df <- Titanic \%>\%
  dplyr::as_tibble() \%>\%
  dplyr::mutate(Survived = factor(Survived, c("No", "Yes")))

df \%>\%
  glm(
    Survived ~ Class + Age + Sex,
    data = ., weights = .$n, family = binomial,
    contrasts = list(Age = contr.sum, Class = "contr.helmert")
  ) \%>\%
  tidy_and_attach() \%>\%
  tidy_add_contrasts()
}
\seealso{
Other tidy_helpers: 
\code{\link{tidy_add_reference_rows}()},
\code{\link{tidy_add_term_labels}()},
\code{\link{tidy_add_variable_labels}()},
\code{\link{tidy_attach_model}()},
\code{\link{tidy_identify_variables}()},
\code{\link{tidy_plus_plus}()}
}
\concept{tidy_helpers}
