% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_add_n.R
\name{tidy_add_n}
\alias{tidy_add_n}
\title{Add the number of observations}
\usage{
tidy_add_n(x, model = tidy_get_model(x))
}
\arguments{
\item{x}{a tidy tibble}

\item{model}{the corresponding model, if not attached to \code{x}}
}
\description{
\lifecycle{maturing}

Add the number of observations in a new column \code{n} taking into account
interaction terms and the different types of contrasts.
}
\details{
For binomial and multinomial logistic models, will also return
the number of events (\code{nevent}).

For Poisson models, will return the number of events (\code{nevent}) and
exposure time defined with \code{\link[stats:offset]{stats::offset()}} (\code{exposure}).

For Cox models (\code{\link[survival:coxph]{survival::coxph()}}), will return the number of
events (\code{nevent}) and exposure time (\code{exposure}).

The total number of observations (\code{N}), of events (\code{Nevent}) and of
exposure time (\code{Exposure}) are stored as attributes of the returned tibble.
}
\examples{
lm(Petal.Length ~ ., data = iris) \%>\%
  tidy_and_attach() \%>\%
  tidy_add_n()

lm(Petal.Length ~ ., data = iris, contrasts = list(Species = contr.sum)) \%>\%
  tidy_and_attach() \%>\%
  tidy_add_n()

lm(Petal.Length ~ ., data = iris, contrasts = list(Species = contr.poly)) \%>\%
  tidy_and_attach() \%>\%
  tidy_add_n()

lm(Petal.Length ~ poly(Sepal.Length, 2), data = iris) \%>\%
  tidy_and_attach() \%>\%
  tidy_add_n()

df <- Titanic \%>\%
  dplyr::as_tibble() \%>\%
  dplyr::mutate(Survived = factor(Survived, c("No", "Yes")))

df \%>\%
  glm(
    Survived ~ Class + Age + Sex,
    data = ., weights = .$n, family = binomial,
    contrasts = list(Age = contr.sum, Class = "contr.helmert")
  ) \%>\%
  tidy_and_attach() \%>\%
  tidy_add_n()

df \%>\%
  glm(
    Survived ~ Class * (Age : Sex),
    data = ., weights = .$n, family = binomial,
    contrasts = list(Age = contr.sum, Class = "contr.helmert")
  ) \%>\%
  tidy_and_attach() \%>\%
  tidy_add_n()

glm(response ~ age + grade * trt, gtsummary::trial, family = poisson) \%>\%
  tidy_and_attach() \%>\%
  tidy_add_n()

glm(
  response ~ trt * grade + offset(log(ttdeath)),
  gtsummary::trial,
  family = poisson
) \%>\%
  tidy_and_attach() \%>\%
  tidy_add_n()
}
\seealso{
Other tidy_helpers: 
\code{\link{tidy_add_coefficients_type}()},
\code{\link{tidy_add_contrasts}()},
\code{\link{tidy_add_estimate_to_reference_rows}()},
\code{\link{tidy_add_header_rows}()},
\code{\link{tidy_add_reference_rows}()},
\code{\link{tidy_add_term_labels}()},
\code{\link{tidy_add_variable_labels}()},
\code{\link{tidy_attach_model}()},
\code{\link{tidy_identify_variables}()},
\code{\link{tidy_plus_plus}()},
\code{\link{tidy_remove_intercept}()},
\code{\link{tidy_select_variables}()}
}
\concept{tidy_helpers}
