% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_add_term_labels.R
\name{tidy_add_term_labels}
\alias{tidy_add_term_labels}
\title{Add term labels}
\usage{
tidy_add_term_labels(
  x,
  labels = NULL,
  interaction_sep = " * ",
  model = tidy_get_model(x)
)
}
\arguments{
\item{x}{a tidy tibble}

\item{labels}{an optional named list or named vector of
custom term labels}

\item{interaction_sep}{separator for interaction terms}

\item{model}{the corresponding model, if not attached to \code{x}}
}
\description{
Will add term labels in a \code{var_label} column, based on:
\enumerate{
\item labels provided in \code{labels} argument if provided;
\item factor levels for categorical variables coded with
treatment, SAS or sum contrasts;
\item variable labels when there is only one term per variable;
\item term name otherwise.
}
}
\details{
If the \code{variable_label} column is not yet available in \code{x},
\code{\link[=tidy_add_variable_labels]{tidy_add_variable_labels()}} will be automatically applied.
If the \code{contrasts} column is not yet available in \code{x},
\code{\link[=tidy_add_contrasts]{tidy_add_contrasts()}} will be automatically applied.

It is possible to pass a custom label for any term in \code{labels},
including interaction terms.
}
\examples{
df <- Titanic \%>\%
  dplyr::as_tibble() \%>\%
  dplyr::mutate(Survived = factor(Survived, c("No", "Yes"))) \%>\%
  labelled::set_variable_labels(
    Class = "Passenger's class",
    Sex = "Sex"
  )

df \%>\%
  glm(Survived ~ Class * Age * Sex, data = ., weights = .$n, family = binomial) \%>\%
  tidy_and_attach() \%>\%
  tidy_add_term_labels()
}
\seealso{
Other tidy_helpers: 
\code{\link{tidy_add_contrasts}()},
\code{\link{tidy_add_reference_rows}()},
\code{\link{tidy_add_variable_labels}()},
\code{\link{tidy_attach_model}()},
\code{\link{tidy_identify_variables}()},
\code{\link{tidy_plus_plus}()}
}
\concept{tidy_helpers}
