% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy_all_effects}
\alias{tidy_all_effects}
\title{Marginal Effects Estimation}
\usage{
tidy_all_effects(x, conf.int = TRUE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{a model}

\item{conf.int}{logical indicating whether or not to include a confidence
interval in the tidied output}

\item{conf.level}{the confidence level to use for the confidence interval}

\item{...}{additional parameters passed to \code{effects::allEffects()}}
}
\description{
Use \code{effects::allEffects()} to estimate "marginal effects" and return a
tibble tidied in a way that could be used by \code{broom.helpers}functions.
See \code{vignette("functions-supported-by-effects", package = "effects")} for
a list of supported models.
}
\examples{
df <- Titanic \%>\%
  dplyr::as_tibble() \%>\%
  dplyr::mutate(Survived = factor(Survived, c("No", "Yes")))
mod <- glm(
  Survived ~ Class + Age + Sex,
  data = df, weights = df$n, family = binomial
)
tidy_all_effects(mod)
tidy_plus_plus(mod, tidy_fun = tidy_all_effects)
}
\seealso{
\code{effects::allEffects()}

Other custom_tieders: 
\code{\link{tidy_ggpredict}()},
\code{\link{tidy_margins}()},
\code{\link{tidy_multgee}()},
\code{\link{tidy_parameters}()},
\code{\link{tidy_with_broom_or_parameters}()}
}
\concept{custom_tieders}
