% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_tidiers.R
\name{tidy_marginal_contrasts}
\alias{tidy_marginal_contrasts}
\title{Marginal Contrasts with \code{marginaleffects::avg_comparisons()}}
\usage{
tidy_marginal_contrasts(
  x,
  variables_list = "auto",
  conf.int = TRUE,
  conf.level = 0.95,
  ...
)
}
\arguments{
\item{x}{a model}

\item{variables_list}{a list whose elements will be sequentially passed to
\code{variables} in \code{marginaleffects::avg_comparisons()} (see details below);
alternatively, it could also be the string \code{"auto"} (default) or
\code{"no_interaction"}}

\item{conf.int}{logical indicating whether or not to include a confidence
interval in the tidied output}

\item{conf.level}{the confidence level to use for the confidence interval}

\item{...}{additional parameters passed to
\code{marginaleffects::avg_comparisons()}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Use \code{marginaleffects::avg_comparisons()} to estimate marginal contrasts for
each variable of a model and return a tibble tidied in a way that it could
be used by \code{broom.helpers} functions.
See \code{marginaleffects::avg_comparisons()} for a list of supported models.
}
\details{
Marginal contrasts are obtained by calling, for each variable or combination
of variables, \code{marginaleffects::avg_comparisons()} with the option \code{cross = TRUE}.

Considering a categorical variable named \code{cat}, \code{tidy_marginal_contrasts()}
will call \code{avg_comparisons(model, variables = list(cat = "reference"))}
to obtain average marginal contrasts for this variable.

Considering a continuous variable named \code{cont}, \code{tidy_marginalcontrasts()}
will call \code{avg_comparisons(model, variables = list(cont = 1))}
to obtain average marginal contrasts for an increase of one unit.

By default, \emph{average marginal contrasts} are computed: contrasts are computed
using a counterfactual grid for each value of the variable of interest,
before averaging the results. \emph{Marginal contrasts at the mean} could be
obtained by indicating \code{newdata = "mean"}. Other assumptions are possible,
see the help file of \code{marginaleffects::avg_comparisons()}.

\code{tidy_marginal_contrasts()} will compute marginal contrasts for each
variable or combination of variables, before stacking the results in a unique
tibble. This is why \code{tidy_marginal_contrasts()} has a \code{variables_list}
argument consisting of a list of specifications that will be passed
sequentially to the \code{variables} argument of \code{marginaleffects::avg_comparisons()}.

The helper function \code{variables_to_predict()} could be used to automatically
generate a suitable list to be used with \code{variables_list}.

\code{variables_list}'s default value, \code{"auto"}, calls
\code{variables_to_predict(interactions = TRUE)} while \code{"no_interaction"} is a
shortcut for \code{variables_to_predict(interactions = FALSE)}. \code{"reference"} is
used for categorical variables and \code{1} for continuous variables.

You can also provide custom specifications (see examples).

For more information, see \code{vignette("marginal_tidiers", "broom.helpers")}.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
# Average Marginal Contrasts
df <- Titanic \%>\%
  dplyr::as_tibble() \%>\%
  tidyr::uncount(n) \%>\%
  dplyr::mutate(Survived = factor(Survived, c("No", "Yes")))
mod <- glm(
  Survived ~ Class + Age + Sex,
  data = df, family = binomial
)
tidy_marginal_contrasts(mod)
tidy_plus_plus(mod, tidy_fun = tidy_marginal_contrasts)

mod2 <- lm(Petal.Length ~ poly(Petal.Width, 2) + Species, data = iris)
tidy_marginal_contrasts(mod2)
tidy_marginal_contrasts(
  mod2,
  variables_list = variables_to_predict(
    mod2,
    continuous = 3,
    categorical = "pairwise"
  )
)

# Model with interactions
mod3 <- glm(
  Survived ~ Sex * Age + Class,
  data = df, family = binomial
)
tidy_marginal_contrasts(mod3)
tidy_marginal_contrasts(mod3, "no_interaction")

# Marginal Contrasts at the Mean
tidy_marginal_contrasts(mod, newdata = "mean")
\dontshow{\}) # examplesIf}
}
\seealso{
\code{marginaleffects::avg_comparisons()}, \code{tidy_avg_comparisons()}

Other marginal_tieders: 
\code{\link{tidy_all_effects}()},
\code{\link{tidy_avg_comparisons}()},
\code{\link{tidy_avg_slopes}()},
\code{\link{tidy_ggpredict}()},
\code{\link{tidy_marginal_means}()},
\code{\link{tidy_marginal_predictions}()},
\code{\link{tidy_margins}()}
}
\concept{marginal_tieders}
