% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/marginal_tidiers.R
\name{tidy_margins}
\alias{tidy_margins}
\title{Average Marginal Effects with \code{margins::margins()}}
\usage{
tidy_margins(x, conf.int = TRUE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{a model}

\item{conf.int}{logical indicating whether or not to include a confidence
interval in the tidied output}

\item{conf.level}{the confidence level to use for the confidence interval}

\item{...}{additional parameters passed to \code{margins::margins()}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}
}
\details{
The ``

Use \code{margins::margins()} to estimate average marginal effects (AME) and
return a tibble tidied in a way that it could be used by \code{broom.helpers}
functions. See \code{margins::margins()} for a list of supported models.

By default, \code{margins::margins()} estimate average marginal effects (AME): an
effect is computed for each observed value in the original dataset before
being averaged.

For more information, see \code{vignette("marginal_tidiers", "broom.helpers")}.
}
\note{
When applying \code{margins::margins()}, custom contrasts are ignored.
Treatment contrasts (\code{stats::contr.treatment()}) are applied to all
categorical variables. Interactions are also ignored.
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
df <- Titanic |>
  dplyr::as_tibble() |>
  tidyr::uncount(n) |>
  dplyr::mutate(Survived = factor(Survived, c("No", "Yes")))
mod <- glm(
  Survived ~ Class + Age + Sex,
  data = df, family = binomial
)
tidy_margins(mod)
tidy_plus_plus(mod, tidy_fun = tidy_margins)
\dontshow{\}) # examplesIf}
}
\seealso{
\code{margins::margins()}

Other marginal_tieders: 
\code{\link{tidy_all_effects}()},
\code{\link{tidy_avg_comparisons}()},
\code{\link{tidy_avg_slopes}()},
\code{\link{tidy_ggpredict}()},
\code{\link{tidy_marginal_contrasts}()},
\code{\link{tidy_marginal_means}()},
\code{\link{tidy_marginal_predictions}()}
}
\concept{marginal_tieders}
