% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy_margins}
\alias{tidy_margins}
\title{Average Marginal Effects Estimation}
\usage{
tidy_margins(x, conf.int = TRUE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{a model}

\item{conf.int}{logical indicating whether or not to include a confidence
interval in the tidied output}

\item{conf.level}{the confidence level to use for the confidence interval}

\item{...}{additional parameters passed to \code{margins::margins()}}
}
\description{
Use \code{margins::margins()} to estimate "average marginal effects" and return a
tibble tidied in a way that could be used by \code{broom.helpers}functions. See
\code{margins::margins()} for a list of supported models.
}
\note{
When applying \code{margins::margins()}, custom contrasts are ignored.
Treatment contrasts (\code{stats::contr.treatment()}) are applied to all
categorical variables. Interactions are also ignored.
}
\seealso{
\code{margins::margins()}

Other custom_tieders: 
\code{\link{tidy_all_effects}()},
\code{\link{tidy_ggpredict}()},
\code{\link{tidy_multgee}()},
\code{\link{tidy_parameters}()},
\code{\link{tidy_with_broom_or_parameters}()}
}
\concept{custom_tieders}
