% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidiers.R
\name{tidy_multgee}
\alias{tidy_multgee}
\title{Tidy a \code{multgee} model}
\usage{
tidy_multgee(x, conf.int = TRUE, conf.level = 0.95, ...)
}
\arguments{
\item{x}{a \code{\link[multgee:nomLORgee]{multgee::nomLORgee()}} or a \code{\link[multgee:ordLORgee]{multgee::ordLORgee()}} model}

\item{conf.int}{logical indicating whether or not to include a confidence
interval in the tidied output}

\item{conf.level}{the confidence level to use for the confidence interval}

\item{...}{additional parameters passed to \code{\link[parameters:model_parameters]{parameters::model_parameters()}}}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\seealso{
Other custom_tieders: 
\code{\link{tidy_parameters}()},
\code{\link{tidy_with_broom_or_parameters}()}
}
\concept{custom_tieders}
