% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_select_variables.R
\name{tidy_select_variables}
\alias{tidy_select_variables}
\title{Select variables to keep/drop}
\usage{
tidy_select_variables(
  x,
  keep = NULL,
  drop = NULL,
  model = tidy_get_model(x),
  quiet = FALSE,
  strict = FALSE
)
}
\arguments{
\item{x}{a tidy tibble}

\item{keep}{variables to keep}

\item{drop}{variables to drop}

\item{model}{the corresponding model, if not attached to \code{x}}

\item{quiet}{logical argument whether a message should not be
returned when requesting to keep/drop a variable not
existing in \code{x}}

\item{strict}{logical argument whether an error should be
returned when requesting to keep/drop a variable not
existing in \code{x}}
}
\description{
Will remove unselected variables from the results.
To remove the intercept, use \code{\link[=tidy_remove_intercept]{tidy_remove_intercept()}}.
}
\details{
If the \code{variable} column is not yet available in \code{x},
\code{\link[=tidy_identify_variables]{tidy_identify_variables()}} will be automatically applied.
}
\examples{
res <- Titanic \%>\%
  dplyr::as_tibble() \%>\%
  dplyr::mutate(Survived = factor(Survived)) \%>\%
  glm(Survived ~ Class + Age + Sex, data = ., weights = .$n, family = binomial) \%>\%
  tidy_and_attach() \%>\%
  tidy_identify_variables()

res
res \%>\% tidy_select_variables()
res \%>\% tidy_select_variables(keep = "Class")
res \%>\% tidy_select_variables(drop = "Class")
res \%>\% tidy_select_variables(keep = c("Age", "Sex"))

}
\seealso{
Other tidy_helpers: 
\code{\link{tidy_add_coefficients_type}()},
\code{\link{tidy_add_contrasts}()},
\code{\link{tidy_add_estimate_to_reference_rows}()},
\code{\link{tidy_add_header_rows}()},
\code{\link{tidy_add_reference_rows}()},
\code{\link{tidy_add_term_labels}()},
\code{\link{tidy_add_variable_labels}()},
\code{\link{tidy_attach_model}()},
\code{\link{tidy_identify_variables}()},
\code{\link{tidy_plus_plus}()},
\code{\link{tidy_remove_intercept}()}
}
\concept{tidy_helpers}
